+++ Sources/Networking/Modifiers/Interceptors/AuthorizationTokenInterceptor.swift
@@ -7,65 +7,6 @@
 
 import Foundation
 
-/// Defines the data the Authorization header is going to be containing.
-/// In case of oAuth, it's going to be Bearer token
-public protocol AuthorizationData {
-    var header: String { get }
-}
-
-// MARK: AuthorizationToken
-public struct AuthorizationToken: AuthorizationData {
-    public let accessToken: String
-    public let refreshToken: String
-    public let expiryDate: Date?
-    
-    public var header: String {
-        "Bearer \(accessToken)"
-    }
-}
-
-// MARK: AuthorizationStorageManaging
-public protocol AuthorizationStorageManaging {
-    associatedtype AuthorizationData = AuthorizationToken
-    
-    func save(data: AuthorizationData) async
-    func get() async -> AuthorizationData?
-    func delete(data: AuthorizationData) async
-}
-
-// MARK: AuthorizationInMemoryStorage
-public actor AuthorizationInMemoryStorage: AuthorizationStorageManaging {
-    private var storage: AuthorizationData?
-    
-    public init() { }
-    
-    public func save(data: AuthorizationData) {
-        storage = data
-    }
-    
-    public func delete(data: AuthorizationData) {
-        storage = nil
-    }
-    
-    public func get() -> AuthorizationData? {
-        storage
-    }
-}
-
-// MARK: Needs to be implemented by end user.
-public protocol AuthorizationManaging {
-    var storage: any AuthorizationStorageManaging { get }
-    func authorize(_ urlRequest: URLRequest) async throws -> URLRequest
-    func refreshToken(_ token: String) async throws
-}
-
-extension AuthorizationManaging {
-    var storage: any AuthorizationStorageManaging {
-        // Default storage, there will be a keychain solution as default.
-        AuthorizationInMemoryStorage()
-    }
-}
-
 // MARK: - Defines authentication handling in requests
 public final class AuthorizationTokenInterceptor: RequestInterceptor {
     private var authorizationManager: AuthorizationManaging
@@ -83,7 +24,18 @@ public final class AuthorizationTokenInterceptor: RequestInterceptor {
     }
     
     public func process(_ response: Response, with urlRequest: URLRequest, for endpointRequest: EndpointRequest) async throws -> Response {
+        guard let httpResponse = response.response as? HTTPURLResponse else {
+            throw NetworkError.noStatusCode(response: response)
+        }
         
-        return (Data(), URLResponse())
+        if httpResponse.statusCode == 401 {
+            print("UNAUTHORIZED")
+        }
+        
+        return response
+    }
+    
+    public func process(_ error: Error, for endpointRequest: EndpointRequest) async -> Error {
+        error
     }
 }
